# docs/Swagger.yaml
swagger: '2.0'
info:
  description: |
    AND1 Project
  version: pre-alpha-1.0
  title: Event & Equipment Manager
paths:
...
  /event:
    get:
      summary: returns all non hidden events
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - in: query
        name: equipmentId
        type: integer
      - in: query
        name: from
        type: integer
        description: "DateTime from when events are requested. Format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
      - in: query
        name: until
        type: integer
        description: "DateTime until when events are requested. Format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
      produces:
      - application/json
      responses:
        200:
          description: Array of Events
          schema:
            type: array
            items:
              $ref: '#/definitions/event'
        401:
          description: Unauthorized
    post:
      summary: creates new event
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - in: body
        name: Event
        required: true
        description: Event
        schema:
          $ref: '#/definitions/event'
      produces:
      - text/plain
      responses:
        200:
          description: Created
        401:
          description: Unauthorized
        409:
          description: "Conflict, descriptive error message detailing which equipments are not available in given timeframe."
          schema:
            example: "Equipment 0 (id) \"Yamaha LS9-32\" not available\nEquipment 1 (id) \"Yamaha CW118V\" not available"
  /event/{EventId}:
    get:
      summary: returns event with given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: EventId
        in: path
        description: Id of the event to return
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Event
          schema:
            $ref: '#/definitions/event'
        401:
          description: Unauthorized
        404:
          description: Not Found
    delete:
      summary: deletes event with given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: EventId
        in: path
        description: Id of the event to delete
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        200:
          description: Deleted
        401:
          description: Unauthorized
        404:
          description: Not Found
    put:
      summary: updates event with given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: EventId
        in: path
        description: Id of the event to update
        required: true
        type: integer
      - in: body
        name: Event
        required: true
        description: Event with updated attributes
        schema:
          $ref: '#/definitions/event'
      produces:
      - text/plain
      responses:
        200:
          description: Updated
        401:
          description: Unauthorized
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing which equipments are not available in given timeframe."
          schema:
            example: "Equipment 0 (id) \"Yamaha LS9-32\" not available\nEquipment 1 (id) \"Yamaha CW118V\" not available"
...
definitions:
  event:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Generali Night Run Budapest
      hidden:
        type: boolean
        example: false
        description: There are some autogenerated events like repair periods for equipment, which are only manageable indirectly. These events only show up in the equipment’s reserved history, while being hidden in the events calendar.
      start:
        type: integer
        example: 1492992000000
        description: "DateTime, format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
      end:
        type: integer
        example: 1493078400000
        description: "DateTime, format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
      equipments:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Yamaha LS9-32
...